Cleanup

+++ TheCanalaveLibrary.Core/Tags/CharacterPairingType.cs
@@ -1,6 +1,6 @@
 namespace TheCanalaveLibrary.Core;
 
-public enum CharacterPairingType : byte
+public enum CharacterPairingType : short
 {
     Romantic = 0,
     Platonic = 1
